+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `company iam remove user` command
 - `company iam remove group` command
 - `company iam remove serviceaccount` command
+- `company iam remove group-member` command
 
 ### Changed